requirements.txt
Flask==2.0.1
Flask-SQLAlchemy==3.0.3
Flask-WTF==1.1.1
Pillow>=10.4.0

requests==2.26.0
python-dateutil==2.8.2
azure-cognitiveservices-vision-computervision==0.9.0
azure-cognitiveservices-speech==1.24.2
email-validator==1.3.1
python-dotenv==0.19.0
Werkzeug==2.2.3
openai==0.27.0
